Remove int limits and extend to long limits

+++ src/zorbage/type/data/ComplexFloat64Matrix.java
@@ -49,17 +49,17 @@ public class ComplexFloat64Matrix
 	public void multiply(ComplexFloat64MatrixMember a, ComplexFloat64MatrixMember b, ComplexFloat64MatrixMember c) {
 		if (c == a || c == b) throw new IllegalArgumentException("dangerous matrix multiply definition");
 		if (a.cols() != b.rows()) throw new IllegalArgumentException("incompatible matrix shapes in matrix multiply");
-		int rows = a.rows();
-		int cols = b.cols();
-		int common = a.cols(); 
+		long rows = a.rows();
+		long cols = b.cols();
+		long common = a.cols(); 
 		ComplexFloat64Member sum = new ComplexFloat64Member();
 		ComplexFloat64Member atmp = new ComplexFloat64Member();
 		ComplexFloat64Member btmp = new ComplexFloat64Member();
 		ComplexFloat64Member term = new ComplexFloat64Member();
-		for (int row = 0; row < rows; row++) {
-			for (int col = 0; col < cols; col++) {
+		for (long row = 0; row < rows; row++) {
+			for (long col = 0; col < cols; col++) {
 				g.zero(sum);
-				for (int i = 0; i < common; i++) {
+				for (long i = 0; i < common; i++) {
 					a.v(row, i, atmp);
 					b.v(i, col, btmp);
 					g.multiply(atmp, btmp, term);
@@ -105,8 +105,8 @@ public class ComplexFloat64Matrix
 
 	@Override
 	public void zero(ComplexFloat64MatrixMember a) {
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.setV(row, col, ZERO);
 			}
 		}
@@ -117,8 +117,8 @@ public class ComplexFloat64Matrix
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
 		if (a != b)
 			b.init(a.rows(), a.cols());
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.negate(tmp,tmp);
 				b.setV(row, col, tmp);
@@ -136,8 +136,8 @@ public class ComplexFloat64Matrix
 		ComplexFloat64Member atmp = new ComplexFloat64Member();
 		ComplexFloat64Member btmp = new ComplexFloat64Member();
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, atmp);
 				b.v(row, col, btmp);
 				g.add(atmp, btmp, tmp);
@@ -156,8 +156,8 @@ public class ComplexFloat64Matrix
 		ComplexFloat64Member atmp = new ComplexFloat64Member();
 		ComplexFloat64Member btmp = new ComplexFloat64Member();
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, atmp);
 				b.v(row, col, btmp);
 				g.subtract(atmp, btmp, tmp);
@@ -173,8 +173,8 @@ public class ComplexFloat64Matrix
 		if (a.cols() != b.cols()) return false;
 		ComplexFloat64Member value1 = new ComplexFloat64Member();
 		ComplexFloat64Member value2 = new ComplexFloat64Member();
-		for (int r = 0; r < a.rows(); r++) {
-			for (int c = 0; c < a.cols(); c++) {
+		for (long r = 0; r < a.rows(); r++) {
+			for (long c = 0; c < a.cols(); c++) {
 				a.v(r, c, value1);
 				b.v(r, c, value2);
 				if (g.isNotEqual(value1, value2))
@@ -194,8 +194,8 @@ public class ComplexFloat64Matrix
 		if (from == to) return;
 		to.init(from.rows(), from.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < from.rows(); row++) {
-			for (int col = 0; col < from.cols(); col++) {
+		for (long row = 0; row < from.rows(); row++) {
+			for (long col = 0; col < from.cols(); col++) {
 				from.v(row, col, tmp);
 				to.setV(row, col, tmp);
 			}
@@ -228,8 +228,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundTowardsZero(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -242,8 +242,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundAwayFromZero(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -256,8 +256,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundPositive(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -270,8 +270,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundNegative(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -284,8 +284,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundNearest(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -298,8 +298,8 @@ public class ComplexFloat64Matrix
 		if (a != b)
 			b.init(a.rows(), a.cols());
 		ComplexFloat64Member tmp = new ComplexFloat64Member();
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, tmp);
 				g.roundNearestEven(tmp, tmp);
 				b.setV(row, col, tmp);
@@ -310,8 +310,8 @@ public class ComplexFloat64Matrix
 	@Override
 	public boolean isNaN(ComplexFloat64MatrixMember a) {
 		ComplexFloat64Member value = new ComplexFloat64Member();
-		for (int r = 0; r < a.rows(); r++) {
-			for (int c = 0; c < a.cols(); c++) {
+		for (long r = 0; r < a.rows(); r++) {
+			for (long c = 0; c < a.cols(); c++) {
 				a.v(r, c, value);
 				if (g.isNaN(value))
 					return true;
@@ -323,8 +323,8 @@ public class ComplexFloat64Matrix
 	@Override
 	public boolean isInfinite(ComplexFloat64MatrixMember a) {
 		ComplexFloat64Member value = new ComplexFloat64Member();
-		for (int r = 0; r < a.rows(); r++) {
-			for (int c = 0; c < a.cols(); c++) {
+		for (long r = 0; r < a.rows(); r++) {
+			for (long c = 0; c < a.cols(); c++) {
 				a.v(r, c, value);
 				if (g.isInfinite(value))
 					return true;
@@ -340,8 +340,8 @@ public class ComplexFloat64Matrix
 		if (a != b) {
 			b.init(a.rows(), a.cols());
 		}
-		for (int row = 0; row < a.rows(); row++) {
-			for (int col = 0; col < a.cols(); col++) {
+		for (long row = 0; row < a.rows(); row++) {
+			for (long col = 0; col < a.cols(); col++) {
 				a.v(row, col, atmp);
 				g.conjugate(atmp, btmp);
 				b.setV(row, col, btmp);
@@ -354,8 +354,8 @@ public class ComplexFloat64Matrix
 		ComplexFloat64Member value = new ComplexFloat64Member();
 		if (a == b) throw new IllegalArgumentException("cannot transpose in place");
 		b.init(a.cols(), a.rows());
-		for (int r = 0; r < a.rows(); r++) {
-			for (int c = 0; c < a.cols(); c++) {
+		for (long r = 0; r < a.rows(); r++) {
+			for (long c = 0; c < a.cols(); c++) {
 				a.v(r,  c, value);
 				b.setV(c, r, value);
 			}
@@ -379,7 +379,7 @@ public class ComplexFloat64Matrix
 	public void unity(ComplexFloat64MatrixMember a) {
 		ComplexFloat64Member one = new ComplexFloat64Member(1, 0);
 		zero(a);
-		for (int i = 0; i < Math.min(a.rows(), a.cols()); i++) {
+		for (long i = 0; i < Math.min(a.rows(), a.cols()); i++) {
 			a.setV(i, i, one);
 		}
 	}
